server/tnkp/requirements.txt
fastapi>=0.68.0
sqlalchemy>=1.4.0
jinja2>=3.0.0
python-multipart>=0.0.5
uvicorn>=0.15.0
python-dotenv>=0.19.0
itsdangerous>=2.0.0
matplotlib>=3.0.0  # 新增图表库
